// dashboard/src/utils/hasura-api/generated/schemas/setCustomTypesStep.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { CustomTypes } from './customTypes';
import type { SetCustomTypesStepType } from './setCustomTypesStepType';

/**
 * Replace all custom GraphQL types
 */
export interface SetCustomTypesStep {
  type: SetCustomTypesStepType;
  args: CustomTypes;
}