// dashboard/src/utils/hasura-api/generated/schemas/updateActionStep.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { CreateActionArgs } from './createActionArgs';
import type { UpdateActionStepType } from './updateActionStepType';

/**
 * Update an existing action
 */
export interface UpdateActionStep {
  type: UpdateActionStepType;
  args: CreateActionArgs;
}